builder/bb: move busybox-compiler to pkg.
