+++ pkg/uroot/builder/bb_test.go
@@ -8,7 +8,6 @@ import (
 	"go/ast"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"testing"
 
@@ -57,23 +56,4 @@ func findFile(filemap map[string]*ast.File, basename string) *ast.File {
 		}
 	}
 	return nil
-}
-
-func TestPackageRewriteFile(t *testing.T) {
-	dir, err := ioutil.TempDir("", "u-root")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(dir)
-
-	bin := filepath.Join(dir, "foo")
-	if err := BuildBusybox(golang.Default(), []string{"github.com/u-root/u-root/pkg/uroot/test/foo"}, bin); err != nil {
-		t.Fatal(err)
-	}
-
-	cmd := exec.Command(bin)
-	o, err := cmd.CombinedOutput()
-	if err != nil {
-		t.Fatalf("foo failed: %v %v", string(o), err)
-	}
 }
