+++ tools/makebb/makebb.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/u-root/u-root/pkg/golang"
 	"github.com/u-root/u-root/pkg/uroot"
-	"github.com/u-root/u-root/pkg/uroot/builder"
+	"github.com/u-root/u-root/pkg/uroot/builder/bb"
 )
 
 var outputPath = flag.String("o", "bb", "Path to busybox binary")
@@ -41,7 +41,7 @@ func main() {
 		log.Fatal(err)
 	}
 
-	if err := builder.BuildBusybox(env, pkgs, o); err != nil {
+	if err := bb.BuildBusybox(env, pkgs, o); err != nil {
 		log.Fatal(err)
 	}
 }
